Rename: 프로젝트 관련 service 패키지 정리 (#190)
